+++ src/components/CommitHistory.tsx
@@ -1,5 +1,5 @@
-import type { GitLogEntry, HydratedGitBlobObject } from "~/analyzer/model"
-import { Fragment, useEffect, useId, useLayoutEffect, useState } from "react"
+import type { GitLogEntry, HydratedGitTreeObject } from "~/analyzer/model"
+import { Fragment, useEffect, useId, useMemo, useState } from "react"
 import { dateFormatLong } from "~/util"
 import commitIcon from "~/assets/commit_icon.png"
 import type { AccordionData } from "./accordion/Accordion"
@@ -62,23 +62,53 @@ function CommitDistFragment(props: CommitDistFragProps) {
   )
 }
 
+function calculateCommitsForSubTree(tree: HydratedGitTreeObject) {
+  const commitMap = new Map<string, number>()
+  subTree(tree)
+  function subTree(tree: HydratedGitTreeObject) {
+    for (const child of tree.children) {
+      if (!child) continue
+      if (child.type === "blob") {
+        if (!child.commits) continue
+        for (const commit of child.commits) {
+          commitMap.set(commit.hash, commit.time)
+        }
+      } else if (child.type === "tree") {
+        subTree(child)
+      }
+    }
+  }
+  const commitArr = []
+  for (const [hash, time] of commitMap) commitArr.push({hash, time})
+  commitArr.sort((a, b) => b.time - a.time)
+  return commitArr
+}
+
 export function CommitHistory() {
+  const [totalCommitHashes, setTotalCommitHashes] = useState<string[]>([])
   const [commits, setCommits] = useState<GitLogEntry[] | null>(null)
   const [commitIndex, setCommitIndex] = useState(0)
   const commitIncrement = 5
   const { clickedObject } = useClickedObject()
   const fetcher = useFetcher()
 
-  function requestCommits(index: number) {
-    const hashes = (clickedObject as HydratedGitBlobObject).commitsNoTime?.slice(index, index + commitIncrement) ?? []
-    fetcher.load(`/commits?commits=${hashes.join(",")}`)
+  function requestCommits(index: number, commitHashes?: string[]) {
+    if (commitHashes) fetcher.load(`/commits?commits=${commitHashes.slice(index, index + commitIncrement).join(",")}`)
+    else fetcher.load(`/commits?commits=${totalCommitHashes.slice(index, index + commitIncrement).join(",")}`)
   }
 
   useEffect(() => {
     if (clickedObject) {
+      let commitHashes: string[] = []
+      if (clickedObject.type === "blob") {
+        commitHashes = clickedObject.commits.map(a => a.hash)
+      } else {
+        commitHashes = calculateCommitsForSubTree(clickedObject).map((a) => a.hash)
+      }
+      setTotalCommitHashes(commitHashes)
       setCommits(null)
       setCommitIndex(0)
-      requestCommits(0)
+      requestCommits(0, commitHashes)
     }
   }, [clickedObject])
 
@@ -102,7 +132,12 @@ export function CommitHistory() {
   const commitHistoryExpandId = useId()
   const [collapsed, setCollapsed] = useState<boolean>(false)
 
-  if (!clickedObject || clickedObject.type !== "blob") return null
+  const headerText = useMemo<string>(() => {
+    if (!clickedObject) return ""
+    return `Commit history (${Math.min(totalCommitHashes.length, commitIndex + commitIncrement)} of ${totalCommitHashes.length} shown)`
+  }, [clickedObject, totalCommitHashes, commitIndex])
+
+  if (!clickedObject) return null
 
   if (!commits) {
     return (
@@ -127,8 +162,7 @@ export function CommitHistory() {
       <div className="flex cursor-pointer justify-between hover:opacity-70">
         <label className="label grow" htmlFor={commitHistoryExpandId}>
           <h3 className="font-bold">
-            Commit history ({Math.min(clickedObject.noCommits, commitIndex + commitIncrement)} of{" "}
-            {clickedObject.noCommits} shown)
+            {headerText}
           </h3>
         </label>
         <ChevronButton id={commitHistoryExpandId} open={!collapsed} onClick={() => setCollapsed(!collapsed)} />
@@ -149,7 +183,7 @@ export function CommitHistory() {
             Show more commits
           </span>
         ) : fetcher.state === "idle" ? (
-          commitIndex + commitIncrement < clickedObject.noCommits ? (
+          commitIndex + commitIncrement < totalCommitHashes.length? (
             <span
               onClick={() => setCommitIndex(commitIndex + commitIncrement)}
               className="whitespace-pre text-xs font-medium opacity-70 hover:cursor-pointer"
